feat: add proxy-token metadata
