+++ src/component/admin/apiToken/ApiTokenForm/ApiTokenForm.tsx
@@ -1,6 +1,6 @@
-import { Button } from '@mui/material';
+import { Button, Typography } from '@mui/material';
 import { KeyboardArrowDownOutlined } from '@mui/icons-material';
-import React from 'react';
+import React, { useState } from 'react';
 import { useEnvironments } from 'hooks/api/getters/useEnvironments/useEnvironments';
 import useProjects from 'hooks/api/getters/useProjects/useProjects';
 import GeneralSelect from 'component/common/GeneralSelect/GeneralSelect';
@@ -9,6 +9,10 @@ import { useStyles } from './ApiTokenForm.styles';
 import { SelectAllAutocomplete } from './SelectAllAutocomplete/SelectAllAutocomplete';
 import { ApiTokenFormErrorType } from 'component/admin/apiToken/ApiTokenForm/useApiTokenForm';
 import { ConditionallyRender } from 'component/common/ConditionallyRender/ConditionallyRender';
+import { MultiValueInput } from 'component/common/form/MultiValueInput/MultiValueInput';
+import { MultiValueOutput } from 'component/common/form/MultiValueOutput/MultiValueOutput';
+import { parseParameterStrings } from 'utils/parseParameter';
+import { uniqueValues } from 'utils/uniqueValues';
 interface IApiTokenFormProps {
     username: string;
     type: string;
@@ -23,6 +27,8 @@ interface IApiTokenFormProps {
     errors: { [key: string]: string };
     mode: 'Create' | 'Edit';
     clearErrors: (error?: ApiTokenFormErrorType) => void;
+    metadata: { [key: string]: string[] };
+    setMetadata: React.Dispatch<React.SetStateAction<object>>;
 }
 
 const ApiTokenForm: React.FC<IApiTokenFormProps> = ({
@@ -39,9 +45,14 @@ const ApiTokenForm: React.FC<IApiTokenFormProps> = ({
     handleCancel,
     errors,
     clearErrors,
+    metadata,
+    setMetadata,
 }) => {
+    const [inputValues, setInputValues] = useState('');
+
     const TYPE_CLIENT = 'CLIENT';
     const TYPE_ADMIN = 'ADMIN';
+    const TYPE_PROXY = 'PROXY';
     const { classes: styles } = useStyles();
     const { environments: availableEnvironments } = useEnvironments();
     const { projects: availableProjects } = useProjects();
@@ -62,6 +73,33 @@ const ApiTokenForm: React.FC<IApiTokenFormProps> = ({
         label: environment.name,
         disabled: !environment.enabled,
     }));
+
+    const onAddMetadata = () => {
+        setMetadata(prev => {
+            const newValues = uniqueValues([
+                ...(prev.corsDomains || []),
+                ...parseParameterStrings(inputValues),
+            ]);
+
+            console.log(newValues);
+
+            return {
+                corsDomains: newValues,
+            };
+        });
+        setInputValues('');
+    };
+
+    const onRemoveMetaData = (index: number) => {
+        setMetadata(prev => {
+            const copy = [...prev.corsDomains];
+            copy.splice(index, 1);
+
+            return {
+                corsDomains: copy,
+            };
+        });
+    };
 
     return (
         <form onSubmit={handleSubmit} className={styles.form}>
@@ -110,7 +148,7 @@ const ApiTokenForm: React.FC<IApiTokenFormProps> = ({
                     Which environment should the token have access to?
                 </p>
                 <ConditionallyRender
-                    condition={type === TYPE_CLIENT}
+                    condition={type === TYPE_CLIENT || type === TYPE_PROXY}
                     show={
                         <GeneralSelect
                             options={selectableEnvs.map(env => ({
@@ -137,6 +175,25 @@ const ApiTokenForm: React.FC<IApiTokenFormProps> = ({
                             error={errors?.environments}
                             onFocus={() => clearErrors('environments')}
                         />
+                    }
+                />
+
+                <ConditionallyRender
+                    condition={type === TYPE_PROXY}
+                    show={
+                        <>
+                            <Typography>Define CORS domains</Typography>
+                            <MultiValueInput
+                                setInputValues={setInputValues}
+                                inputValues={inputValues}
+                                addValues={onAddMetadata}
+                                error={errors?.metadata}
+                            />
+                            <MultiValueOutput
+                                values={metadata.corsDomains || []}
+                                removeValue={onRemoveMetaData}
+                            />
+                        </>
                     }
                 />
             </div>
