+++ src/utils/uniqueValues.ts
@@ -1,0 +1,3 @@
+export const uniqueValues = <T>(values: T[]): T[] => {
+    return Array.from(new Set(values));
+};
